Add 03_auto_load_driver.diff.
This patch allows the server to select a driver and automatically create a
Devices section when none is specified in your xorg.conf. It chooses the
driver based on a set of PCI ID's that the driver itself provides in
/usr/share/xserver-xorg/pci. If no driver claims the PCI ID of your
primary video card, then it will choose a fallback based on your system.

Also clean up patch fuzz in the rest of the patch stack.