Fix incorrect test regarding keyboard map.
In the map stored in each keyboard device, the first line refers to minimum
keycode, i.e., the 0th line refers to keycode 8. When not using XKB the wrong
test caused some keys to be interpreted as locks ('m' for instance). The had
to be pressed twice to generate both KeyPress and KeyRelease events.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>