debian/local/xvfb-run: Fix use of deprecated tempfile utility
tempfile is pretty well guaranteed to not be available on other
distributions, and continuing to use it over mktemp just encourages bad
scripting practices.

It was unnecessary here anyway as the parent directory was already
securely created and it was only being used to combine file creation and
variable assignment in one command (by taking advantage of the fact that
tempfile, when given a hardcoded filename, will also print the same
filename back to you). It makes much more sense to simply set the
variable, and `touch` the file separately.