Fix reproducibility problems in the xorg-server-source package
The files in the xorg-server.tar.xz tarball had random order, and the
timestamps of patched and newly created files depended on the time of
the build.  Fix that with the --sort, --mtime and --clamp-mtime
options of tar.

Also ensure that the files in the tarball have sane permissions.
There is still a problem if the builder uses a restrictive umask, as
scripts may end up not world-executable, but for usual umask values
like 022 or 002 as on the buildds the xorg-server.tar.xz should always
come out identically.

This also means that xorg-server should build fully reproducible now.